# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Browser4 WebDriver-Compatible API
  description: |
    A WebDriver-compatible HTTP API with selector-first extensions for browser automation.
    This API provides session management, navigation, element interaction, and event handling
    capabilities.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8182
    description: Local development server
...
paths:
...
  /session/{sessionId}:
    get:
      tags:
        - session
      summary: Get session details
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - session
      summary: Delete a session
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/url:
    post:
      tags:
        - navigation
      summary: Navigate to a URL
      operationId: navigate
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUrlRequest'
      responses:
        '200':
          description: Navigation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - navigation
      summary: Get current URL
      operationId: getCurrentUrl
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Current URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/documentUri:
    get:
      tags:
        - navigation
      summary: Get document URI
      operationId: getDocumentUri
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Document URI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/baseUri:
    get:
      tags:
        - navigation
      summary: Get base URI
      operationId: getBaseUri
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Base URI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/reload:
    post:
      tags:
        - navigation
      summary: Reload current page
      operationId: reload
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Page reloaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/back:
    post:
      tags:
        - navigation
      summary: Go back in browser history
      operationId: back
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Navigated back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/forward:
    post:
      tags:
        - navigation
      summary: Go forward in browser history
      operationId: forward
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Navigated forward
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/title:
    get:
      tags:
        - navigation
      summary: Get page title
      operationId: getTitle
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Page title
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/bringToFront:
    post:
      tags:
        - navigation
      summary: Bring window to front
      operationId: bringToFront
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Window brought to front
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session identifier
...
  schemas:
...
    SessionDetails:
      type: object
      properties:
        value:
          type: object
          properties:
            sessionId:
              type: string
            url:
              type: string
            status:
              type: string
              enum: [active, paused, stopped]
            capabilities:
              type: object
              additionalProperties: true
    WebDriverResponse:
      type: object
      properties:
        value:
          description: Response value (can be null for success)
    ErrorResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            error:
              type: string
            message:
              type: string
            stacktrace:
              type: string
    SetUrlRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: URL to navigate to
    UrlResponse:
      type: object
      properties:
        value:
          type: string
          description: Current URL
...
    TextResponse:
      type: object
      properties:
        value:
          type: string
          description: Element text content
...