# pulsar-rest/src/main/kotlin/ai/platon/pulsar/rest/openapi/controller/NavigationController.kt
package ai.platon.pulsar.rest.openapi.controller

import ai.platon.pulsar.rest.openapi.dto.SetUrlRequest
import ai.platon.pulsar.rest.openapi.dto.WebDriverResponse
import ai.platon.pulsar.rest.openapi.service.SessionManager
import jakarta.servlet.http.HttpServletResponse
import org.slf4j.LoggerFactory
import org.springframework.beans.factory.annotation.Value
import org.springframework.boot.autoconfigure.condition.ConditionalOnBean
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity
import org.springframework.web.bind.annotation.*

/**
 * Controller for page navigation operations.
 */
@RestController
@CrossOrigin
@RequestMapping(
    "/session/{sessionId}",
    produces = [MediaType.APPLICATION_JSON_VALUE]
)
@ConditionalOnBean(SessionManager::class)
class NavigationController(
    private val sessionManager: SessionManager,
    @param:Value($$"${pulsar.stub.mode:false}")
    private val stubMode: Boolean = false
) {
    private val logger = LoggerFactory.getLogger(NavigationController::class.java)

    /**
     * Navigates to a URL.
     */
    @PostMapping("/url", consumes = [MediaType.APPLICATION_JSON_VALUE])
    suspend fun navigate(
        @PathVariable sessionId: String,
        @RequestBody request: SetUrlRequest,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} navigating to: {}", sessionId, request.url)
        ControllerUtils.addRequestId(response)

        val session = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        if (stubMode) {
            sessionManager.setSessionUrl(sessionId, request.url)
            return ResponseEntity.ok(WebDriverResponse<Any?>(value = null))
        }

        try {
            // Serialize WebDriver operations using mutex to prevent parallel execution
            session.withLock { driver.navigate(request.url) }
            sessionManager.setSessionUrl(sessionId, request.url)
        } catch (e: Exception) {
            logger.error("Error navigating to URL: {}", e.message, e)
            return ControllerUtils.errorResponse("navigation error", "Failed to navigate: ${e.message}")
        }

        return ResponseEntity.ok(WebDriverResponse<Any?>(value = null))
    }

    /**
     * Gets the current URL.
     */
    @GetMapping("/url")
    fun getCurrentUrl(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} getting current URL", sessionId)
        ControllerUtils.addRequestId(response)

        val session = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        // Return the stored URL from session
        val url = session.url ?: "about:blank"

        return ResponseEntity.ok(WebDriverResponse(value = url))
    }

    /**
     * Gets the document URI.
     */
    @GetMapping("/documentUri")
    fun getDocumentUri(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} getting document URI", sessionId)
        ControllerUtils.addRequestId(response)

        val session = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        // For now, return the current URL (in real implementation, this could be different from URL)
        val uri = session.url ?: "about:blank"

        return ResponseEntity.ok(WebDriverResponse(value = uri))
    }

    /**
     * Gets the base URI.
     */
    @GetMapping("/baseUri")
    fun getBaseUri(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} getting base URI", sessionId)
        ControllerUtils.addRequestId(response)

        val session = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        // Extract base URI from current URL (protocol + host)
        val baseUri = session.url?.let { url ->
            try {
                val uri = java.net.URI(url)
                "${uri.scheme}://${uri.host}${if (uri.port > 0 && uri.port != 80 && uri.port != 443) ":${uri.port}" else ""}"
            } catch (e: Exception) {
                url
            }
        } ?: "about:blank"

        return ResponseEntity.ok(WebDriverResponse(value = baseUri))
    }

    /**
     * Reloads the current page.
     */
    @PostMapping("/reload")
    suspend fun reload(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} reloading page", sessionId)
        ControllerUtils.addRequestId(response)

        val managed = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        return try {
            managed.withLock {
                driver.reload()
            }
            ResponseEntity.ok(WebDriverResponse<Any?>(value = null))
        } catch (e: Exception) {
            logger.error("Reload failed | sessionId={} | {}", sessionId, e.message)
            ControllerUtils.errorResponse("navigation error", e.message ?: "Failed to reload")
        }
    }

    /**
     * Navigates back in browser history.
     */
    @PostMapping("/back")
    suspend fun goBack(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} going back", sessionId)
        ControllerUtils.addRequestId(response)

        val managed = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        return try {
            managed.withLock {
                driver.goBack()
            }
            ResponseEntity.ok(WebDriverResponse<Any?>(value = null))
        } catch (e: Exception) {
            logger.error("Go back failed | sessionId={} | {}", sessionId, e.message)
            ControllerUtils.errorResponse("navigation error", e.message ?: "Failed to go back")
        }
    }

    /**
     * Navigates forward in browser history.
     */
    @PostMapping("/forward")
    suspend fun goForward(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} going forward", sessionId)
        ControllerUtils.addRequestId(response)

        val managed = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        return try {
            managed.withLock {
                driver.goForward()
            }
            ResponseEntity.ok(WebDriverResponse<Any?>(value = null))
        } catch (e: Exception) {
            logger.error("Go forward failed | sessionId={} | {}", sessionId, e.message)
            ControllerUtils.errorResponse("navigation error", e.message ?: "Failed to go forward")
        }
    }

    /**
     * Gets the page title.
     */
    @GetMapping("/title")
    suspend fun getTitle(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} getting title", sessionId)
        ControllerUtils.addRequestId(response)

        val managed = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        return try {
            val title = managed.withLock {
                driver.title()
            }
            ResponseEntity.ok(WebDriverResponse(value = title))
        } catch (e: Exception) {
            logger.error("Get title failed | sessionId={} | {}", sessionId, e.message)
            ControllerUtils.errorResponse("webdriver error", e.message ?: "Failed to get title")
        }
    }

    /**
     * Brings the browser window to the front.
     */
    @PostMapping("/bringToFront")
    suspend fun bringToFront(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} bringing window to front", sessionId)
        ControllerUtils.addRequestId(response)

        val managed = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        return try {
            managed.withLock {
                driver.bringToFront()
            }
            ResponseEntity.ok(WebDriverResponse<Any?>(value = null))
        } catch (e: Exception) {
            logger.error("Bring to front failed | sessionId={} | {}", sessionId, e.message)
            ControllerUtils.errorResponse("webdriver error", e.message ?: "Failed to bring to front")
        }
    }

    /**
     * Gets the page source (HTML).
     */
    @GetMapping("/source")
    suspend fun getPageSource(
        @PathVariable sessionId: String,
        response: HttpServletResponse
    ): ResponseEntity<Any> {
        logger.debug("Session {} getting page source", sessionId)
        ControllerUtils.addRequestId(response)

        val managed = sessionManager.getSession(sessionId)
            ?: return ControllerUtils.notFound("session not found", "No active session with id $sessionId")

        return try {
            val source = managed.withLock {
                driver.pageSource() ?: ""
            }
            ResponseEntity.ok(WebDriverResponse(value = source))
        } catch (e: Exception) {
            logger.error("Get page source failed | sessionId={} | {}", sessionId, e.message)
            ControllerUtils.errorResponse("webdriver error", e.message ?: "Failed to get page source")
        }
    }
}